+++ Source_Files/Input/joystick_sdl.cpp
@@ -165,7 +165,7 @@ void joystick_buttons_become_keypresses(Uint8* ioKeyMap) {
     return;
 }
 
-int process_joystick_axes(int flags) {
+uint64_t process_joystick_axes(uint64_t flags) {
     if (!joystick_active)
         return flags;
 	if (active_instances.empty())
